infra/analytics/refresh_analytics/requirements.txt
psycopg2-binary==2.9.3
boto3